+++ matchms/filtering/__init__.py
@@ -1,4 +1,6 @@
+from .add_adduct import add_adduct
 from .add_compound_name import add_compound_name
+from .add_fingerprint import add_fingerprint
 from .add_losses import add_losses
 from .add_parent_mass import add_parent_mass
 from .add_precursor_mz import add_precursor_mz
@@ -28,7 +30,9 @@ from .SpeciesString import SpeciesString
 
 
 __all__ = [
+    "add_adduct",
     "add_compound_name",
+    "add_fingerprint",
     "add_losses",
     "add_parent_mass",
     "add_precursor_mz",