+++ matchms/similarity/__init__.py
@@ -2,6 +2,7 @@
 from .CosineGreedy import CosineGreedy
 from .CosineGreedyNumba import CosineGreedyNumba
 from .CosineHungarian import CosineHungarian
+from .FingerprintSimilarityParallel import FingerprintSimilarityParallel
 from .IntersectMz import IntersectMz
 from .ModifiedCosine import ModifiedCosine
 
@@ -10,6 +11,7 @@ __all__ = [
     "CosineGreedy",
     "CosineGreedyNumba",
     "CosineHungarian",
+    "FingerprintSimilarityParallel",
     "IntersectMz",
     "ModifiedCosine",
 ]